Q1_MultiAgent_RAG/requirements.txt
langchain>=0.2.0
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
chromadb>=0.4.0
pypdf>=4.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
groq>=0.4.0